Move generated DMA types to dedicated module
This follows the example from other modules, limits the wildcard
re-export to that one module, and makes the `channels` module much
lighter.
